Add files via upload
The main objective of this project, is to be able to build a machine learning model, that can predict the popularity of a song on Spotify, based on various features of the song.

We began this project by finding a dataset collected from the Spotify Web API, that contains contains 169k songs from the year 1921 to year 2020.

The dataset contained 19 columns of different features of the song, including one column for "Popularity". This column contains continuous data, and scores the songs in a range of 0-100, based on their popularity.

Therefore, we defined the "Popularity" as our target variable, and for our projet, we tested different Tree-Based Regression algorithms, to find an optimal model that will be most likely to predict the popularity of a song.